Switch to hatch, drop python 2 support
