+++ azure-template-publish-job.yml
@@ -41,14 +41,11 @@ jobs:
     #
     - download: current
       artifact: ${{ format('sw_{0}_{1}', parameters.tox,  parameters.os) }}
-
-    - download: current
-      artifact: ${{ format('sw_u_{0}_{1}', parameters.tox,  parameters.os) }}
     #
     - script: ls -lR $(Pipeline.Workspace)/${{ format('sw_{0}_{1}', parameters.tox, parameters.os) }}
       displayName: "Listing of downloaded artifacts"
     #
-    - script: python -m pip install --upgrade pip twine setuptools wheel
+    - script: python -m pip install --upgrade pip twine wheel
       displayName: 'Install packaging tools'
     #
     - task: DownloadSecureFile@1
@@ -62,7 +59,3 @@ jobs:
         displayName: "Linux upload sdist and source wheel to PyPi ..."
         continueOnError: true
     #
-    - ${{ if startsWith(parameters.os, 'macos') }}:
-      - script: twine upload --verbose --skip-existing -r pypi  --config-file $(pypicred.secureFilePath) $(Pipeline.Workspace)/${{ format('sw_{0}_{1}', parameters.tox, parameters.os) }}/*
-        displayName: "Mac upload sdist and binary wheel to PyPi ..."
-        continueOnError: true
