+++ azure-template-tox-job.yml
@@ -175,23 +175,14 @@ jobs:
     #  Build artifacts if this is a test target (i.e. labeled as py##)
     #
     - ${{ if startsWith(parameters.tox, 'py') }}:
-      - script: pip install --upgrade pip twine setuptools wheel
+      - script: pip install --upgrade pip twine build wheel
         displayName: "Acquire build tools"
-      - script: python setup.py  sdist --dist-dir "$(System.DefaultWorkingDirectory)/dist"
-        displayName: "Build source dist"
-      - script: python setup.py  bdist_wheel --dist-dir "$(System.DefaultWorkingDirectory)/dist"
-        displayName: "Build wheel"
-    #
-      - script: python setup.py  sdist --dist-dir "$(System.DefaultWorkingDirectory)/udist"
-        displayName: "Build source dist"
-      #
+      - script: python -m build --outdir"$(System.DefaultWorkingDirectory)/dist"
+        displayName: "Build source distributions"
       # Check the install artifacts
-      - script: ls -lR "$(System.DefaultWorkingDirectory)/dist" "$(System.DefaultWorkingDirectory)/udist"
+      - script: ls -lR "$(System.DefaultWorkingDirectory)/dist"
         displayName: "Listing of installed software"
       #
       - publish: $(System.DefaultWorkingDirectory)/dist
         artifact: ${{ format('sw_{0}_{1}', parameters.tox, parameters.os) }}
-      #
-      - publish: $(System.DefaultWorkingDirectory)/udist
-        artifact: ${{ format('sw_u_{0}_{1}', parameters.tox, parameters.os) }}
       #
