--- setup.cfg
@@ -1,7 +1,0 @@
-[bdist_wheel]
-# use py2.py3 tag for pure-python dist:
-universal=1
-
-[metadata]
-description-file = README.md
-
