--- setup.py
@@ -1,64 +1,0 @@
-# File: setup.py
-# Date: 3-Oct-2018
-#
-# Update:
-#
-import re
-
-from setuptools import find_packages
-from setuptools import setup
-
-packages = []
-thisPackage = "wwpdb.utils.markdown_wrapper"
-
-with open("wwpdb/utils/markdown_wrapper/__init__.py", "r") as fd:
-    version = re.search(r'^__version__\s*=\s*[\'"]([^\'"]*)[\'"]', fd.read(), re.MULTILINE).group(1)
-
-if not version:
-    raise RuntimeError("Cannot find version information")
-
-setup(
-    name=thisPackage,
-    version=version,
-    description="wwPDB markdown rendering",
-    long_description="See:  README.md",
-    author="Ezra Peisach",
-    author_email="ezra.peisach@rcsb.org",
-    url="https://github.com/rcsb/py-wwpdb_utils_markdown_wrapper",
-    #
-    license="Apache 2.0",
-    classifiers=[
-        "Development Status :: 3 - Alpha",
-        # 'Development Status :: 5 - Production/Stable',
-        "Intended Audience :: Developers",
-        "Natural Language :: English",
-        "License :: OSI Approved :: Apache Software License",
-        "Programming Language :: Python",
-        "Programming Language :: Python :: 2.7",
-        "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
-    ],
-    #
-    # Code highlighting is better with pygments
-    install_requires=["graphviz", "Markdown~=3.0", "beautifulsoup4", "pygments"],
-    packages=find_packages(exclude=["wwpdb.utils.tests-markdown_wrapper", "mock-data", "tests.*"]),
-    package_data={
-        # If any package contains *.md or *.rst ...  files, include them:
-        "": ["*.md", "*.rst", "*.txt", "*.cfg"],
-    },
-    #
-    # These basic tests require no database services -
-    test_suite="wwpdb.utils.tests-markdown_wrapper",
-    tests_require=["tox"],
-    #
-    # Not configured ...
-    extras_require={
-        "dev": ["check-manifest"],
-        "test": ["coverage"],
-    },
-    # Added for
-    command_options={"build_sphinx": {"project": ("setup.py", thisPackage), "version": ("setup.py", version), "release": ("setup.py", version)}},
-    # This setting for namespace package support -
-    zip_safe=False,
-)
