+++ tests/ImporterTests.py
@@ -7,13 +7,17 @@
 
 import unittest
 
-from wwpdb.utils.markdown_wrapper.render_markdown import markdown2html, addMermaid, getSettings  # noqa: F401 pylint: disable=unused-import
+from wwpdb.utils.markdown_wrapper.render_markdown import (  # noqa: F401 pylint: disable=unused-import
+    addMermaid,
+    getSettings,
+    markdown2html,
+)
 
 
 class ImportTests(unittest.TestCase):
     def testInstantiate(self):
         """Tests simple instantiation"""
-        pass  # pylint: disable=unnecessary-pass
+        # pylint: disable=unnecessary-pass
 
 
 if __name__ == "__main__":
