+++ tests/export.ref
@@ -340,11 +340,95 @@ l_conf --&gt; c_wiz4
     Graphviz manual
    </a>
   </p>
-  <!---- mock svg --->
+  <?xml version="1.0" encoding="UTF-8" standalone="no"?>
+  <!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN"
+ "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">
+  <!-- Generated by graphviz version 12.2.1 (20241206.2353)
+ -->
+  <!-- Title: pet&#45;shop Pages: 1 -->
+  <svg height="44pt" viewbox="0.00 0.00 152.00 44.00" width="152pt" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+   <g class="graph" id="graph0" transform="scale(1 1) rotate(0) translate(4 40)">
+    <title>
+     pet-shop
+    </title>
+    <polygon fill="white" points="-4,4 -4,-40 148,-40 148,4 -4,4" stroke="none">
+    </polygon>
+    <!-- parrot -->
+    <g class="node" id="node1">
+     <title>
+      parrot
+     </title>
+     <text font-family="Times,serif" font-size="14.00" text-anchor="middle" x="27" y="-12.95">
+      parrot
+     </text>
+    </g>
+    <!-- dead -->
+    <g class="node" id="node2">
+     <title>
+      dead
+     </title>
+     <text font-family="Times,serif" font-size="14.00" text-anchor="middle" x="117" y="-12.95">
+      dead
+     </text>
+    </g>
+    <!-- parrot&#45;&gt;dead -->
+    <g class="edge" id="edge1">
+     <title>
+      parrot-&gt;dead
+     </title>
+     <path d="M53.93,-18C58.76,-18 63.95,-18 69.19,-18" fill="none" stroke="black">
+     </path>
+     <polygon fill="black" points="89.04,-18 69.04,-27 80.26,-18 69.04,-18 69.04,-18 69.04,-18 80.26,-18 69.04,-9 89.04,-18" stroke="black">
+     </polygon>
+    </g>
+   </g>
+  </svg>
   <p>
    and another example:
   </p>
-  <!---- mock svg --->
+  <?xml version="1.0" encoding="UTF-8" standalone="no"?>
+  <!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN"
+ "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">
+  <!-- Generated by graphviz version 12.2.1 (20241206.2353)
+ -->
+  <!-- Title: pet&#45;shop Pages: 1 -->
+  <svg height="44pt" viewbox="0.00 0.00 152.00 44.00" width="152pt" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+   <g class="graph" id="graph0" transform="scale(1 1) rotate(0) translate(4 40)">
+    <title>
+     pet-shop
+    </title>
+    <polygon fill="white" points="-4,4 -4,-40 148,-40 148,4 -4,4" stroke="none">
+    </polygon>
+    <!-- parrot -->
+    <g class="node" id="node1">
+     <title>
+      parrot
+     </title>
+     <text font-family="Times,serif" font-size="14.00" text-anchor="middle" x="27" y="-12.95">
+      parrot
+     </text>
+    </g>
+    <!-- dead -->
+    <g class="node" id="node2">
+     <title>
+      dead
+     </title>
+     <text font-family="Times,serif" font-size="14.00" text-anchor="middle" x="117" y="-12.95">
+      dead
+     </text>
+    </g>
+    <!-- parrot&#45;&gt;dead -->
+    <g class="edge" id="edge1">
+     <title>
+      parrot-&gt;dead
+     </title>
+     <path d="M53.93,-18C58.76,-18 63.95,-18 69.19,-18" fill="none" stroke="black">
+     </path>
+     <polygon fill="black" points="89.04,-18 69.04,-27 80.26,-18 69.04,-18 69.04,-18 69.04,-18 80.26,-18 69.04,-9 89.04,-18" stroke="black">
+     </polygon>
+    </g>
+   </g>
+  </svg>
   <h4 id="a-complex-example">
    a Complex example
   </h4>
@@ -354,6 +438,48 @@ l_conf --&gt; c_wiz4
     Graphviz Gallery
    </a>
   </p>
-  <!---- mock svg --->
+  <?xml version="1.0" encoding="UTF-8" standalone="no"?>
+  <!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN"
+ "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">
+  <!-- Generated by graphviz version 12.2.1 (20241206.2353)
+ -->
+  <!-- Title: pet&#45;shop Pages: 1 -->
+  <svg height="44pt" viewbox="0.00 0.00 152.00 44.00" width="152pt" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+   <g class="graph" id="graph0" transform="scale(1 1) rotate(0) translate(4 40)">
+    <title>
+     pet-shop
+    </title>
+    <polygon fill="white" points="-4,4 -4,-40 148,-40 148,4 -4,4" stroke="none">
+    </polygon>
+    <!-- parrot -->
+    <g class="node" id="node1">
+     <title>
+      parrot
+     </title>
+     <text font-family="Times,serif" font-size="14.00" text-anchor="middle" x="27" y="-12.95">
+      parrot
+     </text>
+    </g>
+    <!-- dead -->
+    <g class="node" id="node2">
+     <title>
+      dead
+     </title>
+     <text font-family="Times,serif" font-size="14.00" text-anchor="middle" x="117" y="-12.95">
+      dead
+     </text>
+    </g>
+    <!-- parrot&#45;&gt;dead -->
+    <g class="edge" id="edge1">
+     <title>
+      parrot-&gt;dead
+     </title>
+     <path d="M53.93,-18C58.76,-18 63.95,-18 69.19,-18" fill="none" stroke="black">
+     </path>
+     <polygon fill="black" points="89.04,-18 69.04,-27 80.26,-18 69.04,-18 69.04,-18 69.04,-18 80.26,-18 69.04,-9 89.04,-18" stroke="black">
+     </polygon>
+    </g>
+   </g>
+  </svg>
  </body>
 </html>
