+++ tox.ini
@@ -7,25 +7,25 @@
 test_pattern = "*Tests.py"
 #
 # Source paths (unquoted and space separated list of files/directories) for linting and format checks
-source_paths = wwpdb/utils/markdown_wrapper wwpdb/utils/tests-markdown_wrapper setup.py
+source_paths = wwpdb/utils/markdown_wrapper tests
 #
 # Start directory path for test discovery
 # Each path must reference valid directory that is searchable by python3.9 (i.e. contains __init__.py)
 #  ** It is harmless to point to paths containing no tests.
 #
-test_path_1 = "wwpdb/utils/tests-markdown_wrapper"
+test_path_1 = "tests"
 # These are placeholders valid source directories without tests files
 test_path_2 = "wwpdb/utils/markdown_wrapper"
 test_path_3 = "wwpdb/utils/markdown_wrapper"
 test_path_4 = "wwpdb/utils/markdown_wrapper"
 #
 # Comma separate list of directories/patterns for which test coverage will be evaluated
-coverage_source_paths = "wwpdb/utils/markdown_wrapper,wwpdb/utils/tests-markdown_wrapper"
-coverage_exclude_paths = "wwpdb/__init__.py,wwpdb/utils/tests-markdown_wrapper/test-output/*"
+coverage_source_paths = "wwpdb/utils/markdown_wrapper,wwpdb/utils/tests"
+coverage_exclude_paths = "wwpdb/__init__.py,tests/test-output/*"
 coverage_cutoff = 65
 #
 # Comma list of directories to exclude from flake
-flake_exclude_paths = --exclude wwpdb/utils/tests-markdown_wrapper/test-output
+flake_exclude_paths = --exclude tests/test-output
 
 ## --------------------------------------------------------------------------
 ##  ----------   No project specific setting beyond this point --------------
@@ -37,50 +37,34 @@ flake_exclude_paths = --exclude wwpdb/utils/tests-markdown_wrapper/test-output
 ##
 [tox]
 # The complete list of supported test environments to setup and invoke
-envlist = format_pep8-{py39}, lint_pylint-{py39}, format_black-{py39}, py{39,27}, test_coverage-{py39}
+envlist = format_pep8-{py39}, lint_pylint-{py39}, format_black-{py39}, py{39}, test_coverage-{py39}
 #
 minversion = 3.4.0
 skip_missing_interpreters = true
 skipsdist = false
 
 [testenv]
-passenv = CONFIG_SUPPORT_TOKEN_ENV MOCKEXE
-whitelist_externals = echo
+passenv = MOCKEXE
+allowlist_externals = echo
 # deps = echo
 commands =
     echo "Starting default tests in testenv"
+basepython = py313: python3.13
+         py312: python3.12
+         py311: python3.11
+         py310: python3.10
+         py39: python3.9
+         py38: python3.8	 
 
-[testenv:py27]
+
+[testenv:py3{8,9,10,11,12,13}]
 description = 'Run unit tests (unittest runner) using {envpython}'
-whitelist_externals = echo
+allowlist_externals = echo
 platform=
        macos: darwin
        linux: linux
-basepython =  py27: python2.7
 skip_install = false
 recreate = true
-alwayscopy=true
-usedevelop=true
-deps = -r requirements.txt
-commands =
-    echo "Starting {envname}"
-    {envpython} -V
-    {envpython} -m unittest discover -v --start-directory {[local_settings]test_path_1} --pattern "{[local_settings]test_pattern}"
-    {envpython} -m unittest discover -v --start-directory {[local_settings]test_path_2} --pattern "{[local_settings]test_pattern}"
-    {envpython} -m unittest discover -v --start-directory {[local_settings]test_path_3} --pattern "{[local_settings]test_pattern}"
-    {envpython} -m unittest discover -v --start-directory {[local_settings]test_path_4} --pattern "{[local_settings]test_pattern}"
-    echo "Completed {envname}"
-
-[testenv:py39]
-description = 'Run unit tests (unittest runner) using {envpython}'
-whitelist_externals = echo
-platform=
-       macos: darwin
-       linux: linux
-basepython = py39: python3.9
-skip_install = false
-recreate = true
-alwayscopy=true
 usedevelop=true
 deps = -r requirements.txt
 commands =
@@ -98,7 +82,7 @@ description = 'Run selected PEP8 compliance checks (flake8)'
 platform=
        macos: darwin
        linux: linux
-whitelist_externals = echo
+allowlist_externals = echo
 basepython = py39: python3.9
 deps =
     flake8
@@ -116,7 +100,7 @@ description = 'Run linting compliance checks (pylint)'
 platform=
        macos: darwin
        linux: linux
-whitelist_externals = echo
+allowlist_externals = echo
 basepython3.9 = py39: python3.9
 
 deps =
@@ -133,7 +117,7 @@ description = 'Run format compliance checks (black)'
 platform=
        macos: darwin
        linux: linux
-whitelist_externals = echo
+allowlist_externals = echo
 basepython3.9 = py39: python3.9
 deps =
     black>=19.10b0
@@ -146,15 +130,13 @@ commands =
     echo "Completed {envname}"
 
 #
-[testenv:test_coverage-py39]
+[testenv:test_coverage-py3{8,9,10,11,12,13}]
 description = 'Run test coverage analysis'
 platform=
        macos: darwin
        linux: linux
-whitelist_externals = echo
-basepython = py39: python3.9
+allowlist_externals = echo
 recreate = true
-alwayscopy=true
 usedevelop=true
 deps =
     coverage
